BB-3468: Make a template for the popup window (frontend)  - restore changes after backmerge revert

+++ src/Oro/Bundle/WindowsBundle/Resources/public/js/widget/dialog-widget.js
@@ -140,7 +140,7 @@ define(function(require) {
                 return;
             }
             dialogManager.remove(this);
-            if (this.model) {
+            if (this.model && !this.options.preventToRemoveModel) {
                 this.model.destroy({
                     error: _.bind(function(model, xhr) {
                         // Suppress error if it's 404 response and not debug mode
